Refactor new classes and interfaces into separate files.
